# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DefenSYS API
  description: |
    Code-derived OpenAPI specification for the DefenSYS Django REST backend.
    Base path prefix: /api/

    See DEFENSYS_REAL_SYSTEM_FLOW.md for permissions, data persistence, and gap analysis.

    Demo-fill and seed-demo prototype endpoints have been removed from production builds.
  version: 1.0.0
  contact:
    name: DefenSYS
servers:
  - url: http://localhost:8000/api
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /teams/weekly-progress/:
    get:
      tags: [weekly-progress]
      summary: List weekly progress reports
      responses:
        '200':
          description: Reports list
    post:
      tags: [weekly-progress]
      summary: Create report (team leader)
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
      responses:
        '201':
          description: Created
  /teams/weekly-progress/{pk}/:
    get:
      tags: [weekly-progress]
      summary: Get report
      parameters:
        - name: pk
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Report detail
    put:
      tags: [weekly-progress]
      summary: Update report (team leader)
      parameters:
        - name: pk
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Updated
    delete:
      tags: [weekly-progress]
      summary: Delete report (team leader)
      parameters:
        - name: pk
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: Deleted
...
  /teams/documents/{document_id}/:
    get:
      tags: [documents]
      summary: Document metadata
      parameters:
        - name: document_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Metadata
    delete:
      tags: [documents]
      summary: Delete document
      parameters:
        - name: document_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: Deleted
...